295.1. Redux Setup{yarn add redux react-redux --save}(Section 21: Handling Authentication with React)
In this video we're going to start installing redux and setting it up inside of our project so we can

eventually get some action creators and to reduce or to reflect whether or not these are signed in once

we reflect the authentications state of the user inside of a redux story.

It's going to make building the rest of our application a lot easier.

OK.

So to get started we're going to flip on over to our terminal.

We're going to stop my running server and we're going to install a couple of dependencies here still

doing NPM install dash dash save.

And we're going to get Rydex and re-act dash redux as well.

OK.

So I'm going to let those install while that goes.

All flip back over to my code editor now.

Right now I'm going to close the Google off file.

I still have this file right here with my client ID.

I'm going to close this down.

I'm not going to save that file because we already copied the client id over to our component and then

I'll close down my jazz file as wel