297.1. Building the Auth Reducer(Section 21: Handling Authentication with React)
We're now calling the appropriate action creator any time that our state changes according to the API

library in this video we're now going to start working on douceur that's going to record whether or

not the user is signed in at any given point in time.

Now remember the entire idea behind recording the authentication status inside of our redock store is

to make sure that the authentication status is available to other components were other screens inside

of application if we had stuck with the approach we have before where only the Google component knew

whether or not the user was signed in that it would be really challenging for other components to understand

whether or not they should treat the user as being signed in or signed out.

OK.

So are going to close down all the files I have open and then I'll open up my reducers directory and

inside there are going to make a new file called off reducer.

Yes.

And then inside of here we'll put together some boiler plate or a reducer.

So I'm going to have a function that gets called with some state objects.

Going to default this to be an object as opposed to just a boolean flag and also have some action property

and they might be curious why are we using a default value of a object as opposed to just like true

false or null.

Well in truth we're going to eventually have one other property related to authentication and you'll

see what that isn't just a little bit.

So in addition to having just some like true or false or no value to indicate whether or not the user

signed in we're going to have some other piece of data inside of here as well.

And the easiest way to arrange all this data is to make use of an object.

Now this object right here I expect to eventually have some property inside of it called something like

is signed in.

And we probably want to start off with that property be an asset to know.

So in order to default the argument of state in this case or in order to initialize it I'm actually

going to create a initialiser up here.

It's going to make a new variable called initial state and then inside if you're going to put the initial

value for my state object.

So I WANT IT TO HAVE A is signed in property and I wanted to be defaulted or initialized to No.