297.2. Building the Auth Reducer(Section 21: Handling Authentication with React)
So now when we default the value state right here rather than assigning it a completely empty object

I'm going to assign it initial state.

Remember the idea here with the default argument when our redux application first boots up are reduced

or it gets called one time to initialize it.

If this producer gets called with a first argument value of undefined then state will be set equal to

whatever value we put right here.

So in order to get a initial value that has an object with the property inside of it of null I'm going

to just to find the object up here ahead of time and then set the default value right there.

Just you know I have a variable name here that is all capitalized to make sure that it's really clear

to other engineers that this is supposed to be a true constant object and we should never change any

of the values inside of it.

We usually indicate that by making use of this capitalized syntax so the capitalized syntax means do

not try to modify this object under any circumstance whatsoever.

OK so now inside of a producer we're going to add a switch over our action type and we'll have a case

of what was it.

Let's check our actions file we have a case of sign in and a case of sign out so sign in and a case

of sign out and I also put in my default case in which I will return my state object.

So now in the case of sign in I want to take my state object I want to update or modify a property inside

of it.

Remember anytime you want to modify a property we're going to use that spread syntax that dot dot dot.

So I going to create an entirely new object here.

I'm going to take all the values out of my state argument and put it into that new object and then I'm

going to also update the is signed in property as well.

And I'm going to update that to be true.

It's now it can do the same thing with sign out as well.

So I will return dot dot dot state is signed out and this case we are no longer signed in because we

just signed out.

So we're going to update is signed.

It's signed in to be false.

Like so OK.

That's pretty much it.

So I'm going to save this producer and then going to flip over to the reducers index file.

I will import that reduce or at the top.

So Auth. reducer from author reducer and then I'm going to assign that to a key on my state object of

art reducer.

So

all right so this looks pretty good.

Now once again I'm going to flip on over to my browser and just do a quick refresh try clicking the

button and make sure I don't see any error message or anything like that.

So looks pretty good to me.

So let's take a quick pause right here.

The last thing we're going to do is kind of close the loop here so to speak.

So we're now recording our authentication status inside of a redux store where we need to make sure

that our Google component either shows the sign in or sign out button depending upon whether or not

these are signed in.

The last thing we do is communicate that state back over to the Google off component.

So quick break and I'll see you in just a minute.