305.1. Connecting Redux Form(Section 23: Handling Forms with Redux Form)
In this video we're going to start to wire up redux form to our project.

We already installed the library two videos ago.

So make sure that you have installed it as well and started up your server again.

Once that's all done I'm going to flip on over to my code editor and we're going to open up our reducers

index file.

So reducers index yes.

Remember when we just discussed redux form a little bit ago I had told you that redux form is going

to do a tremendous amount of stuff for us automatically.

One of the things that redux form is going to do for us is manage all of our form data inside of our

redux store in order to do so we have to wire up producer to our store.

But this is going to be a producer that has been created for us by the redux form library.

So we are not creating this producer.

It has already already been created.

We just installed it into our project when we installed redux form.

So all we have to do is pull that reduce or out of the library and hook it up to our combined reducers

call right here.

So to do so I'm going to add in an import statement at the top.

I'm going to say import reducer from redux Dasch form and then we have to assign this to a very particular

key inside of the combined reducers call.

We are required to wire it up to a key of form.

So I'll say form is coming from reducer.

Now remember any time that we have named exports from a library or a file that we'd write we have to

write out the exact name of that named expert right here.

So we have to import this thing as a variable called reducer.

However a lot of times you might have a lot of different reducers inside of your project.

And so to just have a variable inside of your called reducer is a little bit confusing.

So as a little helpful thing this is absolutely not required at all this is just a helpful thing.

Very frequently we will to use a little bit of syntax here to rename this named export.

When we pull it into this file so we can write optionally reducer as form reducer like so.